SC-15596 Add separate scripts
