+++ docs/synthetics/puppeteer-scripts/load-web-page.js
@@ -1,0 +1,8 @@
+async function testPage(page) {
+    // visit the home page
+    await page.goto('https://www.apple.com/');
+    // take screenshot of the home page
+    await page.screenshot({ path: 'screenshot.jpg' });
+}
+
+module.exports = testPage;
