+++ docs/synthetics/puppeteer-scripts/text-search.js
@@ -1,0 +1,7 @@
+async function testPage(page) {
+    await page.goto('https://sematext.com/');
+    await page.waitForXPath('//*[contains(., "Troubleshooting Just Got Easier")]');
+    await page.screenshot({ path: 'screenshot.jpg' });
+}
+
+module.exports = testPage;
